+++ bin/log.js
@@ -2,7 +2,7 @@ var through = require('through2')
 var pump = require('pump')
 var ndjson = require('ndjson')
 var debug = require('debug')('bin/versions')
-var openDat = require('../lib/open-dat.js')
+var Dat = require('..')
 var abort = require('../lib/abort.js')
 var usage = require('../lib/usage.js')('log.txt')
 
@@ -18,19 +18,14 @@ function handleLog (args) {
     return usage()
   }
 
-  openDat(args, function ready (err, db) {
-    if (err) abort(err, args)
-    handleReadStream(db)
-  })
+  var dat = Dat(args)
 
-  function handleReadStream (db) {
-    var formatter
-    if (args.json) formatter = ndjson.serialize()
-    else formatter = through.obj(format)
-    pump(db.createChangesStream(args), formatter, process.stdout, function done (err) {
-      if (err) abort(err, args, 'dat: err in versions')
-    })
-  }
+  var formatter
+  if (args.json) formatter = ndjson.serialize()
+  else formatter = through.obj(format)
+  pump(dat.db.createChangesStream(args), formatter, process.stdout, function done (err) {
+    if (err) abort(err, args, 'dat: err in versions')
+  })
 
   function format (obj, enc, next) {
     // its the root node, lets not print it out
